requirements.txt
streamlit
pandas
altair==5.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.5.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
compress-pickle==2.1.0
contourpy==1.3.0
folium==0.17.0 
gitdb==4.0.11
GitPython==3.1.43
streamlit==1.38.0
hydralit==1.0.14
hydralit_components==1.0.10
idna==3.8
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leafmap==0.38.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.6.2
numpy==1.25.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.0
protobuf==5.28.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit_folium==0.23.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.34.0
watchdog==4.0.2
xyzservices==2024.9.0
